Separate sharable part and Twitter specific part
